backend-getCar-getDelete-getReservation

+++ modules/parking-and-bike/controllers/deleteReservation.js
@@ -4,10 +4,28 @@ const deleteReservation = async (req, res) => {
     const { reservation_id } = req.params;
 
     try {
+        //find id and to get parking_slot_id
+        const reservation = await prisma.parking_reservation.findUnique({
+            where: {
+                id: parseInt(reservation_id),
+            },
+        });
+
+        if (!reservation) {
+            return res.status(404).json({ error: `Reservation with ID ${reservation_id} does not exist.` });
+        }
+
+        const { parking_slot_id } = reservation;
+        
         const deletedReservation = await prisma.parking_reservation.delete({
             where: {
                 id: parseInt(reservation_id)
             },
+        });
+
+        await prisma.parking_slot.update({
+            where: { id: parking_slot_id },
+            data: { status: true }
         });
 
         res.json({
@@ -21,3 +39,5 @@ const deleteReservation = async (req, res) => {
 };
 
 export { deleteReservation };
+
+// to delete by Admin
